+++ packages/eslint-config-react-typescript/index.js
@@ -1,16 +1,20 @@
 module.exports = {
-  extends: ['@winwangqi/eslint-config-react', '@winwangqi/eslint-config-typescript'].map(require.resolve),
+  extends: [
+    '@winwangqi/eslint-config-react',
+    '@winwangqi/eslint-config-typescript',
+  ].map(require.resolve),
   rules: {
-    "import/extensions": [
-      "error",
-      "ignorePackages",
+    'import/extensions': [
+      'error',
+      'ignorePackages',
       {
-        "js": "never",
-        "jsx": "never",
-        "ts": "never",
-        "tsx": "never"
-      }
+        js: 'never',
+        jsx: 'never',
+        ts: 'never',
+        tsx: 'never',
+      },
     ],
+    'no-useless-constructor': 'off',
   },
   settings: {
     'import/resolver': {