+++ packages/eslint-config-react-typescript/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @winwangqi/eslint-config-react-typescript
 
+## 1.1.4
+
+### Patch Changes
+
+- change import and class rules
+- Updated dependencies
+  - @winwangqi/eslint-config-react@1.1.4
+  - @winwangqi/eslint-config-typescript@1.1.4
+
 ## 1.1.3
 
 ### Patch Changes